requirements.txt
asgiref==3.5.0
backports.zoneinfo==0.2.1
certifi==2021.10.8
charset-normalizer==2.0.11
confusable-homoglyphs==3.2.0
coreapi==2.3.3
coreschema==0.0.4
defusedxml==0.7.1
diff-match-patch==20200713
Django==4.0.1
django-admin-interface==0.18.4
django-archive==0.2.0
django-colorfield==0.6.3
django-filter==21.1
django-flat-responsive==2.0
django-flat-theme==1.1.4
django-import-export==2.7.1
django-location-field==2.1.0
django-registration==3.2
django-registration-redux==2.9
django-tinymce==3.4.0
django-versatileimagefield==2.2
djangorestframework==3.13.1
djangorestframework-simplejwt==5.0.0
docutils==0.18.1
drf-yasg==1.20.0
et-xmlfile==1.1.0
idna==3.3
inflection==0.5.1
itypes==1.2.0
Jinja2==3.0.3
MarkupPy==1.14
MarkupSafe==2.0.1
odfpy==1.4.1
openpyxl==3.0.9
packaging==21.3
Pillow==9.0.0
psycopg2-binary==2.9.3
PyJWT==2.3.0
pyparsing==3.0.7
python-decouple==3.5
python-magic==0.4.25
pytz==2021.3
PyYAML==6.0
requests==2.27.1
ruamel.yaml==0.17.20
ruamel.yaml.clib==0.2.6
six==1.16.0
sqlparse==0.4.2
tablib==3.2.0
uritemplate==4.1.1
urllib3==1.26.8
xlrd==2.0.1
xlwt==1.3.0
